Add StopTest handling
According to:

    http://fitnesse.org/FitNesse.UserGuide.SliM.ExceptionHandling

When an exception is raised with 'StopTest' in its name, the test should
be aborted (in contrast with the default Slim behavior, which is to
continue anyway). This feature was not apparently not implemented in
Rubyslim; this commit implements it. Whenever a generic exception is
caught by the statement executor, if its name contains 'StopTest', then
the string 'STOP' is prepended to the usual exception message. This
'STOP' string is then checked by the list executor, and the test is
aborted immediately if it's found.